+++ package.json
@@ -7,7 +7,7 @@
 		"sheetloaf": "./dist/index.js"
 	},
 	"scripts": {
-		"build": "esbuild src/*.ts --outdir=dist --bundle --platform=node --target=node10.4 --watch --sourcemap=inline",
+		"build": "esbuild src/index.ts --outdir=dist --bundle --platform=node --target=node10.4",
 		"test": "mocha",
 		"test2": "node . \"test/samples/styles/**/*.scss\" --dir \"test/samples/render/\" --load-path \"test/samples/lib\" --style expanded --base test/samples/styles/ --poll 500 --use autoprefixer,postcss-custom-properties --watch",
 		"test3": "cat test/samples/styles/file.scss | node . --style compressed --use autoprefixer --load-path test/samples/styles > test/samples/render/file.css",
@@ -34,14 +34,15 @@
 	},
 	"homepage": "https://github.com/cheddarbread/sheetloaf#readme",
 	"devDependencies": {
+		"@types/node": "^16.11.9",
 		"autoprefixer": "^10.4.0",
 		"esbuild": "^0.13.15",
 		"lodash": "^4.17.21",
 		"mocha": "^9.1.3",
 		"postcss": "^8.3.11",
 		"postcss-custom-properties": "^12.0.0",
-		"typescript": "^4.5.2",
-		"@types/node": "^16.11.9"
+		"ts-node": "^10.4.0",
+		"typescript": "^4.5.2"
 	},
 	"peerDependencies": {
 		"postcss": "^8.2.0"
